Create File-Size Histograms
This was the final project for the course. It pulls the file size of each file in a given directory and generates a histogram that both groups files by size and displays what percentage a specific range of file sizes makes up the directory. It was designed by my partner, James McGuire, and myself.