Customized C version of the Unix command "cp"
This is a version of the Unix command "cp" written in C. It takes a source file and copies it line by line to either a preexisting file and a new file created by the program. This source code is initialized into an executable "mycp" with the line "g++ -o mycp mycp.c".