Basic Unix shell written in C++
This is a shell that can read command lines and execute certain commands. It was written in such a way that it executes basic scripts, including script.txt, the test script. It can run a command in the background while another command is executing. It also ignores lines that begin with '#' because it recognizes those lines and comments.